# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CopilotResponse": {
        "properties": {
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "title": "Missing Skills",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "resume_suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Resume Suggestions",
            "type": "array"
          },
          "why_match": {
            "items": {
              "type": "string"
            },
            "title": "Why Match",
            "type": "array"
          }
        },
        "required": [
          "match_score",
          "why_match",
          "missing_skills",
          "resume_suggestions",
          "model_version"
        ],
        "title": "CopilotResponse",
        "type": "object"
      },
      "CoverLetterResponse": {
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "grounding_notes": {
            "items": {
              "type": "string"
            },
            "title": "Grounding Notes",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "requires_human_review": {
            "title": "Requires Human Review",
            "type": "boolean"
          }
        },
        "required": [
          "body",
          "grounding_notes",
          "requires_human_review",
          "model_version"
        ],
        "title": "CoverLetterResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityDetailResponse": {
        "properties": {
          "application_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Status"
          },
          "apply_url": {
            "title": "Apply Url",
            "type": "string"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "description_text": {
            "title": "Description Text",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_remote": {
            "title": "Is Remote",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "location",
          "is_remote",
          "apply_url",
          "source_kind",
          "skills",
          "updated_at",
          "description_text"
        ],
        "title": "OpportunityDetailResponse",
        "type": "object"
      },
      "OpportunityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpportunityResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "OpportunityListResponse",
        "type": "object"
      },
...
      "RecommendationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecommendationListResponse",
        "type": "object"
      },
      "RemotePreference": {
        "enum": [
          "any",
          "remote_only",
          "hybrid_or_onsite"
        ],
        "title": "RemotePreference",
        "type": "string"
      },
      "SeekerCriteriaRequest": {
        "properties": {
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Locations",
            "type": "array"
          },
          "remote_preference": {
            "$ref": "#/components/schemas/RemotePreference",
            "default": "any"
          },
          "resume_text": {
            "default": "",
            "title": "Resume Text",
            "type": "string"
          },
          "salary_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "INR",
            "title": "Salary Currency"
          },
          "salary_expectation_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Expectation Min"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "years_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "title": "SeekerCriteriaRequest",
        "type": "object"
      },
      "SeekerCriteriaResponse": {
        "properties": {
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Locations",
            "type": "array"
          },
          "remote_preference": {
            "title": "Remote Preference",
            "type": "string"
          },
          "resume_text": {
            "title": "Resume Text",
            "type": "string"
          },
          "salary_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Currency"
          },
          "salary_expectation_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Expectation Min"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "years_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "required": [
          "resume_text",
          "skills",
          "years_experience",
          "preferred_locations",
          "salary_expectation_min",
          "salary_currency",
          "remote_preference"
        ],
        "title": "SeekerCriteriaResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "career-os",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/criteria": {
      "get": {
        "operationId": "getSeekerCriteria",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeekerCriteriaResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get seeker ranking criteria",
        "tags": [
          "jobs"
        ]
      },
      "put": {
        "operationId": "upsertSeekerCriteria",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekerCriteriaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeekerCriteriaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert seeker ranking criteria and recompute matches",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/me/recommendations": {
      "get": {
        "operationId": "listRecommendations",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Highly relevant job recommendations",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/opportunities": {
      "get": {
        "operationId": "searchOpportunities",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "remote_only",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Remote Only"
            }
          },
          {
            "in": "query",
            "name": "company",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search active opportunities",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "getOpportunity",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity detail",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/opportunities/{opportunity_id}/copilot": {
      "get": {
        "operationId": "getOpportunityCopilot",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Career Copilot advice for an opportunity",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/opportunities/{opportunity_id}/cover-letter": {
      "post": {
        "operationId": "generateCoverLetter",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverLetterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a grounded cover-letter draft (human review required)",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}